Mouse pos is now ints instead of floats

+++ Rush-Engine/src/Rush/core/Input.cpp
@@ -48,6 +48,6 @@ bool Input::KeyReleaseHandle(KeyboardReleaseEvent &e){
 
 bool Input::KeyDown(int keycode) { return s_Instance->m_KeyStates[keycode]; }
 bool Input::MousePressed(int button) { return s_Instance->m_MouseButtonStates[button]; }
-std::pair<float, float> Input::MousePos() { return s_Instance->m_MousePos; }
+std::pair<int, int> Input::MousePos() { return s_Instance->m_MousePos; }
 
 }
